chore: reverted two hooks meant for use-query
